+++ RELEASE_NOTES.md
@@ -2,7 +2,10 @@
 
 ## 0.0.2-dev - 2026-08-19
 
-- no changes
+### Dependencies
+
+- Add support for `authoritative` mode and resource destruction (PR #3 by
+  @dependabot[bot])
 
 ## 0.0.1 - 2026-08-19